refactor(api): move controller method
updateUserAccountFromRecoveryDemand
