+++ api/tests/identity-access-management/unit/application/account-recovery/account-recovery.controller.test.js
@@ -1,0 +1,45 @@
+import { accountRecoveryController } from '../../../../../src/identity-access-management/application/account-recovery/account-recovery.controller.js';
+import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
+import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
+import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
+
+describe('Unit | Identity Access Management | Application | Controller | account-recovery', function () {
+  describe('#updateUserForAccountRecovery', function () {
+    it('calls updateUserForAccountRecovery usecase and return 204', async function () {
+      // given
+      const user = domainBuilder.buildUser({ id: 1 });
+      const temporaryKey = 'validTemporaryKey';
+      const domainTransaction = Symbol();
+
+      const request = {
+        params: {
+          id: user.id,
+        },
+        payload: {
+          data: {
+            attributes: {
+              password: user.password,
+              'temporary-key': temporaryKey,
+            },
+          },
+        },
+      };
+
+      sinon.stub(usecases, 'updateUserForAccountRecovery').resolves();
+      sinon.stub(DomainTransaction, 'execute').callsFake((lambda) => {
+        return lambda(domainTransaction);
+      });
+
+      // when
+      const response = await accountRecoveryController.updateUserAccountFromRecoveryDemand(request, hFake);
+
+      // then
+      expect(usecases.updateUserForAccountRecovery).calledWithMatch({
+        password: user.password,
+        temporaryKey,
+        domainTransaction,
+      });
+      expect(response.statusCode).to.equal(204);
+    });
+  });
+});
