+++ api/tests/unit/application/account-recovery/account-recovery-controller_test.js
@@ -1,7 +1,6 @@
 import { accountRecoveryController } from '../../../../lib/application/account-recovery/account-recovery-controller.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
-import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
+import { expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Controller | account-recovery-controller', function () {
   describe('#sendEmailForAccountRecovery', function () {
@@ -81,45 +80,6 @@ describe('Unit | Controller | account-recovery-controller', function () {
       expect(
         studentInformationForAccountRecoverySerializerStub.serializeAccountRecovery,
       ).to.have.been.calledWithExactly(studentInformation);
-    });
-  });
-
-  describe('#updateUserForAccountRecovery', function () {
-    it('should call updateUserForAccountRecovery usecase and return 204', async function () {
-      // given
-      const user = domainBuilder.buildUser({ id: 1 });
-      const temporaryKey = 'validTemporaryKey';
-      const domainTransaction = Symbol();
-
-      const request = {
-        params: {
-          id: user.id,
-        },
-        payload: {
-          data: {
-            attributes: {
-              password: user.password,
-              'temporary-key': temporaryKey,
-            },
-          },
-        },
-      };
-
-      sinon.stub(usecases, 'updateUserForAccountRecovery').resolves();
-      sinon.stub(DomainTransaction, 'execute').callsFake((lambda) => {
-        return lambda(domainTransaction);
-      });
-
-      // when
-      const response = await accountRecoveryController.updateUserAccountFromRecoveryDemand(request, hFake);
-
-      // then
-      expect(usecases.updateUserForAccountRecovery).calledWithMatch({
-        password: user.password,
-        temporaryKey,
-        domainTransaction,
-      });
-      expect(response.statusCode).to.equal(204);
     });
   });
 });
